Modified the test structure to follow convention with how the lib directory is going to be laid out.
The Dashboard class itself will call in modules, with each module representing a section of teh Dashboard  API docs page.

Also made a modification to the dashboard_api.rb class requirements to allow it to actually require organizations.rb
correctly.
